Move interface from 'executors' package
