+++ src/main/java/net/hydromatic/sqllogictest/ISqlTestOperation.java
@@ -20,7 +20,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.hydromatic.sqllogictest.executors;
+package net.hydromatic.sqllogictest;
 
 import net.hydromatic.sqllogictest.util.ICastable;
