+++ src/main/java/net/hydromatic/sqllogictest/SltTestFile.java
@@ -22,7 +22,6 @@
  */
 package net.hydromatic.sqllogictest;
 
-import net.hydromatic.sqllogictest.executors.ISqlTestOperation;
 import net.hydromatic.sqllogictest.util.Utilities;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
